+++ ionoscloud/test_constants.go
@@ -1862,6 +1862,7 @@ const clusterCredentialsPasswordAttribute = "password"
 const datacenterResourceName = "datacenter_example"
 const lanResourceName = "lan_example"
 const mongoVersion = "7.0"
+const mongoVersionOld = "6.0"
 
 // Values
 const clusterMaintenanceWindowDayOfTheWeekValue = "Sunday"